add operator.rb
